feat: Implement cmd using cobra
