+++ .github/workflows/build.yml
@@ -6,9 +6,6 @@ jobs:
   build:
     runs-on: ubuntu-latest
     if: github.event.pusher.name != 'dreamkast-cloudnativedays'
-    strategy:
-      matrix:
-        bin: [initdb, serve]
     steps:
       - uses: actions/checkout@v3
 
@@ -31,7 +28,7 @@ jobs:
         id: meta
         uses: docker/metadata-action@v4
         with:
-          images: dreamkast-weaver/${{ matrix.bin }} # TODO ${{ steps.login-ecr.outputs.registry }}/dreamkast-weaver/${{ matrix.bin }} 
+          images: dreamkast-weaver # TODO ${{ steps.login-ecr.outputs.registry }}/dreamkast-weaver
           tags: |
             type=sha,prefix=,format=long
             type=ref,event=branch
@@ -41,7 +38,7 @@ jobs:
         uses: docker/build-push-action@v4
         with:
           context: ./
-          file: ./dockerfiles/${{ matrix.bin }}.dockerfile
+          file: Dockerfile
           builder: ${{ steps.buildx.outputs.name }}
           push: false # TODO true
           provenance: false
@@ -53,7 +50,7 @@ jobs:
       #- name: Run Trivy vulnerability scanner
       #  uses: aquasecurity/trivy-action@master
       #  with:
-      #    image-ref: '${{ steps.login-ecr.outputs.registry }}/dreamkast-weaver/${{ matrix.bin }}:${{ github.sha }}'
+      #    image-ref: '${{ steps.login-ecr.outputs.registry }}/dreamkast-weaver:${{ github.sha }}'
       #    format: 'table'
       #    exit-code: '0'
       #    ignore-unfixed: true
