+++ cmd/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "dreamkast-weaver/cmd/dkw"
+
+func main() {
+	dkw.Execute()
+}
