+++ tools/tools.go
@@ -8,4 +8,5 @@ import (
 	_ "github.com/ServiceWeaver/weaver/cmd/weaver"
 	_ "github.com/amacneil/dbmate/v2"
 	_ "github.com/kyleconroy/sqlc/cmd/sqlc"
+	_ "github.com/spf13/cobra-cli"
 )
